+++ CHANGELOG.md
@@ -16,6 +16,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 - Fix checkbox/switch validity status
+- Split Calendar component's `value: Date | Date[]` property into two properties - `value: Date` and `values: Date[]`
+- Replaced Calendar component's `hasHeader` property & `has-header` attribute with `hideHeader` & `hide-header` respectively.
+- Replaced Card component's `outlined` property with `elevated`.
 
 ## 1.0.0 - 2021-11-22